+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
    `Symfony\Component\Serializer\Normalizer\NormalizerInterface` or implement `NormalizerAwareInterface` instead
  * Add service usages list to the `debug:container` command output
  * Add service and alias deprecation message to `debug:container [<name>]` output
+ * Add a `mailer:send` command
 
 6.1
 ---